Moved Filtering, Cell Rendering/Editing into sub sections.
