+++ src/javascript-grid-cell-rendering/index.php
@@ -429,439 +429,6 @@ colDef.cellRenderer = function(params) {
 };
 </pre>
 </p>
-<!-- React from here on -->
-<h2 id="reactCellRendering">
-    <img src="../images/react_large.png" style="width: 60px;"/>
-    React Cell Rendering
-</h2>
-
-<p>
-    It is possible to provide a React cellRenderer for ag-Grid to use. All of the information above is
-    relevant to React cellRenderers. This section explains how to apply this logic to your React component.
-</p>
-
-<p>
-    For examples on React cellRendering, see the
-    <a href="https://github.com/ceolter/ag-grid-react-example">ag-grid-react-example</a> on Github.
-    In the example, both 'Skills' and 'Proficiency' columns use React cellRenderers. The Country column
-    uses a standard ag-Grid cellRenderer, to demonstrate both working side by side.</p>
-</p>
-
-<h3><img src="../images/react_large.png" style="width: 20px;"/> Specifying a React cellRenderer</h3>
-
-<p>
-    If you are using the ag-grid-react component to create the ag-Grid instance,
-    then you will have the option of additionally specifying the cellRenderers
-    as React components.
-</p>
-
-<pre><span class="codeComment">// create your cellRenderer as a React component</span>
-class NameCellRenderer extends React.Component {
-    render() {
-    <span class="codeComment">// put in render logic</span>
-        return &lt;span>{this.props.value}&lt;/span>;
-    }
-}
-
-<span class="codeComment">// then reference the Component in your colDef like this</span>
-colDef = {
-
-    <span class="codeComment">// instead of cellRenderer we use cellRendererFramework</span>
-    cellRendererFramework: NameCellRenderer
-
-    <span class="codeComment">// specify all the other fields as normal</span>
-    headerName: 'Name',
-    field: 'firstName',
-    ...
-}</pre>
-
-<p>
-    By using <i>colDef.cellRendererFramework</i> (instead of <i>colDef.cellRenderer</i>) the grid
-    will know it's a React component, based on the fact that you are using the React version of
-    ag-Grid.
-</p>
-
-<p>
-    This same mechanism can be to use a React Component in the following locations:
-<ul>
-    <li>colDef.cellRenderer<b>Framework</b></li>
-    <li>colDef.floatingCellRenderer<b>Framework</b></li>
-    <li>gridOptions.fullWidthCellRenderer<b>Framework</b></li>
-    <li>gridOptions.groupRowRenderer<b>Framework</b></li>
-    <li>gridOptions.groupRowInnerRenderer<b>Framework</b></li>
-</ul>
-In other words, wherever you specify a normal cellRenderer, you can now specify a React cellRenderer
-in the property of the same name excepting ending 'Framework'. As long as you are using the React ag-Grid component,
-the grid will know the framework to use is React.
-</p>
-
-<h3><img src="../images/react_large.png" style="width: 20px;"/> React Props</h3>
-
-<p>
-    The React component will get the 'cellRenderer Params' as described above as it's React Props.
-    Therefore you can access all the parameters as React Props.
-
-<pre><span class="codeComment">// React cellRenderer Component</span>
-class NameCellRenderer extends React.Component {
-
-    <span class="codeComment">// did you know that React passes props to your component constructor??</span>
-    constructor(props) {
-        super(props);
-        <span class="codeComment">// from here you can access any of the props!</span>
-        console.log('The value is ' + props.value);
-        <span class="codeComment">// we can even call grid API functions, if that was useful</span>
-        props.api.selectAll();
-    }
-
-    render() {
-        <span class="codeComment">// or access props using 'this'</span>
-        return &lt;span>{this.props.value}&lt;/span>;
-    }
-}</pre>
-</p>
-
-<h3><img src="../images/react_large.png" style="width: 20px;"/> React Methods / Lifecycle</h3>
-
-<p>
-    All of the methods in the ICellRenderer interface described above are applicable
-    to the React Component with the following exceptions:
-<ul>
-    <li><i>init()</i> is not used. Instead use the React props passed to your Component.</li>
-    <li><i>destroy()</i> is not used. Instead use the React <i>componentWillUnmount()</i> method for
-        any cleanup you need to do.</li>
-    <li><i>getGui()</i> is not used. Instead do normal React magic in your <i>render()</i> method..</li>
-</ul>
-
-<h3><img src="../images/react_large.png" style="width: 20px;"/> Handling Refresh</h3>
-
-<p>
-    You have the option of handling refresh or not by either providing a <i>refresh()</i> method on
-    your React component or not. If not present, then the grid will destroy your component and create
-    a new one if it tries to refresh the cell. If you do implement it, then it's up to your React
-    components <i>refresh()</i> method to update the state of your component.
-</p>
-
-<!-- Angular from here -->
-<h2 id="ng2CellRendering">
-    <img src="../images/angular2_large.png" style="width: 60px;"/>
-    Angular Cell Rendering
-</h2>
-
-<p>
-    It is possible to provide a Angular cellRenderers for ag-Grid to use. All of the information above is
-    relevant to Angular cellRenderers. This section explains how to apply this logic to your Angular component.
-</p>
-
-<p>
-    For examples on Angular cellRendering, see the
-    <a href="https://github.com/ceolter/ag-grid-ng2-example">ag-grid-ng2-example</a> on Github.
-    Angular Renderers are used on all but the first Grid on this example page (the first grid uses plain JavaScript Renderers)</p>
-</p>
-
-<h3><img src="../images/angular2_large.png" style="width: 20px;"/> Specifying a Angular cellRenderer</h3>
-
-<p>
-    If you are using the ag-grid-ng2 component to create the ag-Grid instance,
-    then you will have the option of additionally specifying the cellRenderers
-    as Angular components.
-
-<h3><img src="../images/angular2_large.png" style="width: 20px;"/> cellRenderers from Angular Components</h3>
-<pre><span class="codeComment">// create your cellRenderer as a Angular component</span>
-@Component({
-    selector: 'square-cell',
-    template: `{{valueSquared()}}`
-})
-class SquareComponent implements AgRendererComponent {
-    private params:any;
-
-    agInit(params:any):void {
-        this.params = params;
-    }
-
-    private valueSquared():number {
-        return this.params.value * this.params.value;
-    }
-}
-<span class="codeComment">// then reference the Component in your colDef like this</span>
-colDef = {
-    {
-        headerName: "Square Component",
-        field: "value",
-        <span class="codeComment">// instead of cellRenderer we use cellRendererFramework</span>
-        cellRendererFramework: SquareComponent
-
-        <span class="codeComment">// specify all the other fields as normal</span>
-        editable:true,
-        colId: "square",
-        width: 200
-    }
-}</pre>
-
-<p>Your Angular components need to implement <code>AgRendererComponent</code>.
-    The ag Framework expects to find the <code>agInit</code> method on the created component, and uses it to supply the cell <code>params</code>.<p>
-
-<p>
-    By using <i>colDef.cellRendererFramework</i> (instead of <i>colDef.cellRenderer</i>) the grid
-    will know it's a Angular component, based on the fact that you are using the Angular version of
-    ag-Grid.
-</p>
-
-<p>
-    This same mechanism can be to use a Angular Component in the following locations:
-<ul>
-    <li>colDef.cellRenderer<b>Framework</b></li>
-    <li>colDef.floatingCellRenderer<b>Framework</b></li>
-    <li>gridOptions.fullWidthCellRenderer<b>Framework</b></li>
-    <li>gridOptions.groupRowRenderer<b>Framework</b></li>
-    <li>gridOptions.groupRowInnerRenderer<b>Framework</b></li>
-</ul>
-In other words, wherever you specify a normal cellRenderer, you can now specify a Angular cellRenderer
-in the property of the same name excepting ending 'Framework'. As long as you are using the Angular ag-Grid component,
-the grid will know the framework to use is Angular.
-</p>
-
-<h3>Example: Rendering using Angular Components</h3>
-<p>
-    Using Angular Components in the Cell Renderers
-</p>
-<show-example example="../ng2-example/index.html?fromDocs=true&example=from-component"
-              jsfile="../ng2-example/app/from-component.component.ts"
-              html="../ng2-example/app/from-component.component.html"></show-example>
-
-<h3><img src="../images/angular2_large.png" style="width: 20px;"/> Angular Methods / Lifecycle</h3>
-
-<p>
-    All of the methods in the ICellRenderer interface described above are applicable
-    to the Angular Component with the following exceptions:
-<ul>
-    <li><i>init()</i> is not used. Instead implement the <code>agInit</code> method (on the <code>AgRendererComponent</code> interface).</li>
-    <li><i>destroy()</i> is not used. Instead implement the Angular<code>OnDestroy</code> interface (<code>ngOnDestroy</code>) for
-        any cleanup you need to do.</li>
-    <li><i>getGui()</i> is not used. Instead do normal Angular magic in your Component via the Angular template.</li>
-</ul>
-
-<h3><img src="../images/angular2_large.png" style="width: 20px;"/> Handling Refresh</h3>
-
-<p>To receive update (for example, after an edit) you should implement the optional <code>refresh</code> method on the <code>AgRendererComponent</code> interface.</p>
-
-<h3>Example: Rendering using more complex Angular Components</h3>
-<p>
-    Using more complex Angular Components in the Cell Renderers
-</p>
-<show-example example="../ng2-example/index.html?fromDocs=true&example=from-rich-component"
-              jsfile="../ng2-example/app/from-rich.component.ts"
-              html="../ng2-example/app/from-rich.component.html"></show-example>
-
-<note>The full <a href="https://github.com/ceolter/ag-grid-ng2-example">ag-grid-ng2-example</a> repo shows many more examples for rendering, including grouped rows, full width renderers
-    and so on, as well as examples on using Angular Components with both CellEditors and Filters</note>
-
-<!-- vuejs from here -->
-<h2 id="vueCellRendering">
-    <img src="../images/vue_large.png" style="width: 60px;"/>
-    VueJS Cell Rendering
-</h2>
-
-<p>
-    It is possible to provide a VueJS cellRenderers for ag-Grid to use. All of the information above is
-    relevant to VueJS cellRenderers. This section explains how to apply this logic to your VueJS component.
-</p>
-
-<p>
-    For examples on VueJS cellRendering, see the
-    <a href="https://github.com/ceolter/ag-grid-vue-example">ag-grid-vue-example</a> on Github.
-    VueJS Renderers are used on all but the first Grid on this example page (the first grid uses plain JavaScript Renderers)</p>
-</p>
-
-<h3><img src="../images/vue_large.png" style="width: 20px;"/> Specifying a VueJS cellRenderer</h3>
-
-<p>
-    If you are using the ag-grid-vue component to create the ag-Grid instance,
-    then you will have the option of additionally specifying the cellRenderers
-    as VueJS components.</p>
-
-    <p>A VueJS component can be defined in a few different ways (please see <a href="/best-vuejs-data-grid#define_component">
-        Defining VueJS Components</a> for all the options), but in this example we're going to define our renderer as a Single File Component:</p>
-
-<pre ng-non-bindable><span class="codeComment">// create your cellRenderer as a VueJS component</span>
-&lt;template&gt;
-    &lt;span class="currency"&gt;{{ params.value | currency('EUR') }}&lt;/span&gt;
-&lt;/template&gt;
-
-&lt;script&gt;
-    import Vue from "vue";
-
-    export default Vue.extend({
-        filters: {
-            currency(value, symbol) {
-                let result = value;
-                if (!isNaN(value)) {
-                    result = value.toFixed(2);
-                }
-                return symbol ? symbol + result : result;
-            }
-        }
-    });
-&lt;/script&gt;
-
-&lt;style scoped&gt;
-    .currency {
-        color: blue;
-    }
-&lt;/style&gt;
-
-<span class="codeComment">// then reference the Component in your colDef like this</span>
-{
-    <span class="codeComment">// instead of cellRenderer we use cellRendererFramework</span>
-    cellRendererFramework: CurrencyComponent,
-
-    <span class="codeComment">// specify all the other fields as normal</span>
-    headerName: "Currency (Filter)",
-    field: "currency",
-    colId: "params",
-    width: 150
-}
-</pre>
-
-<p>The Grid cell's value will be made available implicitly in a data value names <code>params</code>. This value will be available to
-    you from the <code>created</code> VueJS lifecycle hook.</p>
-
-<p>You can think of this as you having defined the following:</p>
-<pre>
-export default {
-    data () {
-        return {
-            params: null
-        }
-    },
-    ...
-</pre>
-
-<p>but you do not need to do this - this is made available to you behind the scenes, and contains the cells value.</p>
-
-<p>
-    By using <i>colDef.cellRendererFramework</i> (instead of <i>colDef.cellRenderer</i>) the grid
-    will know it's a VueJS component, based on the fact that you are using the VueJS version of
-    ag-Grid.
-</p>
-
-<p>
-    This same mechanism can be to use a VueJS Component in the following locations:
-<ul>
-    <li>colDef.cellRenderer<b>Framework</b></li>
-    <li>colDef.floatingCellRenderer<b>Framework</b></li>
-    <li>gridOptions.fullWidthCellRenderer<b>Framework</b></li>
-    <li>gridOptions.groupRowRenderer<b>Framework</b></li>
-    <li>gridOptions.groupRowInnerRenderer<b>Framework</b></li>
-</ul>
-In other words, wherever you specify a normal cellRenderer, you can now specify a VueJS cellRenderer
-in the property of the same name excepting ending 'Framework'. As long as you are using the VueJS ag-Grid component,
-the grid will know the framework to use is VueJS.
-</p>
-
-<h3>Example: Rendering using VueJS Components</h3>
-<p>
-    Using VueJS Components in the Cell Renderers
-</p>
-<show-example url="../vue-examples/#/dynamic"
-              jsfile="../vue-examples/src/dynamic-component-example/DynamicComponentExample.vue"
-              exampleHeight="525px"></show-example>
-
-<h3><img src="../images/vue_large.png" style="width: 20px;"/> VueJS Methods / Lifecycle</h3>
-
-<p>
-    All of the methods in the ICellRenderer interface described above are applicable
-    to the VueJS Component with the following exceptions:
-<ul>
-    <li><i>init()</i> is not used. The cells value is made available implicitly via a data field called <code>params</code>.</li>
-    <li><i>getGui()</i> is not used. Instead do normal VueJS magic in your Component via the VueJS template.</li>
-</ul>
-
-<h3><img src="../images/vue_large.png" style="width: 20px;"/> Handling Refresh</h3>
-
-<p>To receive update (for example, after an edit) you should implement the optional <code>refresh</code> method on the <code>AgRendererComponent</code> interface.</p>
-
-<h3>Example: Rendering using more complex VueJS Components</h3>
-<p>
-    Using more complex VueJS Components in the Cell Renderers
-</p>
-<show-example url="../vue-examples/#/rich-dynamic"
-              jsfile="../vue-examples/src/rich-dynamic-component-example/RichDynamicComponentExample.vue"
-              exampleHeight="525px"></show-example>
-
-<note>The full <a href="https://github.com/ceolter/ag-grid-vue-example">ag-grid-vue-example</a> repo shows many more examples for rendering, including grouped rows, full width renderers
-    and so on, as well as examples on using VueJS Components with both CellEditors and Filters</note>
-
-<!-- Aurelia from here -->
-<h2 id="aureliaCellRendering">
-    <img src="../images/aurelia_large.png" style="width: 60px;"/>
-    Aurelia Cell Rendering
-</h2>
-
-<p>
-    It is possible to provide Aurelia Component cellRenderers for ag-Grid to use, with support for two way binding. All of the information above is
-    relevant to Aurelia cellRenderers. This section explains how to apply this logic to your Aurelia component.
-</p>
-
-<p>
-    For examples on Aurelia cellRendering, see the
-    <a href="https://github.com/ceolter/ag-grid-ng2-example">ag-grid-aurelia-example</a> on Github.</p>
-</p>
-
-<h3><img src="../images/aurelia_large.png" style="width: 20px;"/> Specifying a Aurelia cellRenderer</h3>
-
-<p>
-    If you are using the ag-grid-aurelia component to create the ag-Grid instance,
-    then you will have the option of additionally specifying the cellRenderers
-    as aurelia components. You have two options that are described below:
-<ol>
-    <li>Templates</li>
-    <li>cellRenderer Functions or cellRenderer Components</li>
-</ol>
-
-    We only mention Templates here - using functions or regular components is discussed above in earlier sections.
-</p>
-
-
-<h3><img src="../images/aurelia_large.png" style="width: 20px;"/> cellRenderers from Templates</h3>
-<pre><span class="codeComment">// create your cellRenderer as an Aurelia template</span>
-&lt;ag-grid-aurelia #agGrid style="width: 100%; height: 100%;" class="ag-fresh"
-                 grid-options.bind="gridOptions">
-  &lt;ag-grid-column header-name="Mood" field="mood" width.bind="150" editable.bind="true">
-    &lt;ag-cell-template>
-      &lt;img width="20px" if.bind="params.value === 'Happy'" src="images/smiley.png"/>
-      &lt;img width="20px" if.bind="params.value !== 'Happy'" src="images/smiley-sad.png"/>
-    &lt;/ag-cell-template>
-  &lt;/ag-grid-column>
-&lt;/ag-grid-aurelia>
-</pre>
-
-<p>The advantage of using Templates for your renderers is that you have the ability to use Aurelia's dynamic data binding facilties.</p>
-
-<h3>Example: Rendering using Templates</h3>
-<p>
-    Using Templates in the Cell Renderers - here we're using a Template (<code>ag-cell-template</code>) to render the Mood column depending on the underlying value
-    of the "mood" field.
-</p>
-
-<show-example example="../aurelia-example/#/editor/true"
-              jsfile="../aurelia-example/components/editor-example/editor-example.ts"
-              html="../aurelia-example/components/editor-example/editor-example.html"></show-example>
-
-<h3><img src="../images/aurelia_large.png" style="width: 20px;"/> Aurelia Methods / Lifecycle</h3>
-
-
-<h3>Example: Rendering using regular cellRenderer Components</h3>
-<p>
-    A Rich Grid leveraging regular Cell Renderer Components
-</p>
-
-<show-example example="../aurelia-example/#/richgrid-declarative/true"
-          jsfile="../aurelia-example/components/rich-grid-declarative-example/rich-grid-declarative-example.ts"
-          html="../aurelia-example/components/rich-grid-declarative-example/rich-grid-declarative-example.html"
-          exampleHeight="525px"></show-example>
-
-<br/>
-<note>The full <a href="https://github.com/ceolter/ag-grid-aurelia-example">ag-grid-aurelia-example</a> repo shows many more examples for rendering, including grouped rows, full width renderers and filters.</note>
 
 <?php include '../documentation-main/documentation_footer.php';?>
